Generalized news logic
The newsfeed now reads in an author, source tuple. Every 60 seconds, it now randomly selects from the sources, checks any posts from the last 24 hours, and then verifies if it has seen it before with its tracker table (seen_post_ids.csv). If it has not seen it before, it posts it to the discord channel.